backend/requirements.txt
fastapi
python-multipart
python-dotenv
supabase
uvicorn
